pages fetch data properly

+++ pages/api/protected/questions/delete.js
@@ -1,13 +1,20 @@
 import prisma from "lib/prisma";
 
 export default async function handler(req, res) {
-  const { id } = req.body;
+  if (req.method !== "DELETE") {
+    return res.status(404).json({ error: "Method not allowed" });
+  }
+  try {
+    const { id } = req.body;
 
-  await prisma.question.delete({
-    where: {
-      id: Number(id),
-    },
-  });
+    await prisma.question.delete({
+      where: {
+        id: Number(id),
+      },
+    });
 
-  return res.status(200).json();
+    return res.status(200).json();
+  } catch {
+    return res.status(500).json({ error });
+  }
 }
